+++ src/_/types.ts
@@ -1,4 +1,5 @@
-export type Anchor = `bottom` | `bottom-left` | `bottom-right` | `left` | `top` | `top-left` | `top-right` | `right`;
+export type Anchor =
+    `bottom` | `bottom-left` | `bottom-right` | `center` | `left` | `top` | `top-left` | `top-right` | `right`;
 export type Environment = `debug` | `offline` | `production`;
 export type Media = HTMLImageElement | HTMLVideoElement;
 export type Mode = `contain` | `cover`;