+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Add
+- Add `Anchor` type export to **Angular** components.
+
 ## [0.14.1] - 2022-12-08
 
 ### Fixed